pyproject.toml
[tool.poetry]
name = "coursework-6-skymarket"
version = "0.1.0"
description = ""
authors = ["mosgo <eugeny,ga@gmail.com>"]
readme = "README.md"
packages = [{include = "coursework_6_skymarket"}]

[tool.poetry.dependencies]
python = "^3.11"
drf-spectacular = "^0.26.3"
django-cors-headers = "^4.2.0"
psycopg2 = "^2.9.6"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.2.2"
pytest = "^7.4.0"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
